pythonservice/requirements.txt
astroid==1.6.3
click==6.7
Flask==1.0.2
gunicorn==19.8.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pylint==1.8.4
PyMySQL==0.8.1
python-dotenv==0.8.2
six==1.11.0
Werkzeug==0.14.1
wrapt==1.10.11
